requirements.txt
django==3.0.2
djangorestframework
django-rest-auth